Create helpers folder for utility functions. Refactor into async/await and leverage fs.promises for clearer order of ops
